server/requirements.txt
edgy[postgres]==0.32.5
lilya==0.16.9
bcrypt==4.3.0
fastapi==0.116.1
uvicorn==0.35.0
uvicorn[standard]==0.35.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.1
psycopg2-binary==2.9.10
pgvector==0.4.1
alembic==1.16.4
pydantic==2.11.7
pydantic[email]==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
bs4==0.0.2
click==8.2.1
rich==14.0.0
email-validator==2.2.0
aiosmtplib==4.0.1
jinja2==3.1.6
html2text==2025.4.15
sqlparse==0.5.3
openpyxl==3.1.5
pyexcel-ods3==0.6.1
pandas==2.3.1
openai==1.96.0
dateparser==1.2.2
mistralai==1.9.2
assemblyai==0.42.0
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
email-reply-parser==0.5.12
mail-parser==4.1.4
mail-parser-reply==1.34
